+++ src/main/java/com/efs/sdk/organizationmanager/commons/OrganizationmanagerException.java
@@ -57,6 +57,7 @@ public class OrganizationmanagerException extends Exception {
         GET_SINGLE_SPACE_NOT_FOUND(10021, HttpStatus.NOT_FOUND, "space does not exist."),
         INVALID_NAME(10022, HttpStatus.BAD_REQUEST,
                 "Invalid name provided, be aware that the name must apply to the following regular expression"),
+        INVALID_DATA(10023, HttpStatus.BAD_REQUEST, "Invalid object data."),
         NO_ACCESS_TO_ORGANIZATION(10025, HttpStatus.FORBIDDEN, "You do not have access to get organization"),
         NO_ACCESS_TO_SPACE(10026, HttpStatus.FORBIDDEN, "You do not have access to get space"),
         RENAMING_OBJECT_FORBIDDEN(10027, HttpStatus.FORBIDDEN, "renaming the object is forbidden"),
@@ -75,6 +76,7 @@ public class OrganizationmanagerException extends Exception {
         CONFLICTING_IDS_PROVIDED(20031, HttpStatus.CONFLICT, "conflicting object-ids provided"),
         VALIDATION_ERROR(40000, HttpStatus.BAD_REQUEST, "dto validation error."),
         UNABLE_DELETE_ORGA(40001, HttpStatus.BAD_REQUEST, "could not delete organization."),
+        BAD_REQUEST_PARAM_VALUE(40002, HttpStatus.BAD_REQUEST, "bad request param value."),
 
         UNKNOWN_ERROR(50000, HttpStatus.INTERNAL_SERVER_ERROR, "something unexpected happened."),
         DOWNSTREAM_ERROR(50200, HttpStatus.BAD_GATEWAY, "downstream error: "),